requirements.txt
attrs~=19.3.0
click~=7.1.2
hikari==1.0.1.dev137
